Initiated re-structure of Client Service Request process
