+++ resources/views/client/services/_contactList.blade.php
@@ -8,29 +8,25 @@
         <tbody>
             @if($myContacts) 
                 @foreach($myContacts as $k=>$myContact)
-                        <tr>
-                            <td>
-                                <div class="media">
-                                
-                                    <div class="custom-control custom-radio custom-control-inline">
-                                        <div class="form-group mb-0">
-                                            <input type="radio" id="{{$myContact->id}}" value="{{ $myContact->id }}" name="myContact_id" class="custom-control-input"  />
-                                            <!-- <input type="radio" id="{{$myContact->id}}" name="id" value="{{ $myContact->id }}" {{ ( (isset($myContact->is_default) && intval($myContact->is_default)) ? 'checked=checked' : '') }}> -->
-                                            <input type="hidden" name="state_id" value="{{ $myContact->state_id }}">
-                                            <input type="hidden" name="lga_id" value="{{ $myContact->lga_id }}">
-                                            <input type="hidden" name="town_id" value="{{ $myContact->town_id }}">
-                                            <label class="custom-control-label" for="{{$myContact->id}}">{{$myContact->name ?? ''}}</label>
-                                        </div>
+                    <tr>
+                        <td>
+                            <div class="media">
+                            
+                                <div class="custom-control custom-radio custom-control-inline">
+                                    <div class="form-group mb-0">
+                                        <input type="radio" id="{{$myContact->id}}" value="{{ $myContact->id }}" name="contact_id" class="custom-control-input contact-id" data-town-id="{{ $myContact->town_id }}" data-contact-name="{{ $myContact->name ?? 'Default Contact' }}" data-url="{{ route('client.service-request.validate_service_area', app()->getLocale()) }}" />
+                                        <label class="custom-control-label" for="{{$myContact->id}}">{{$myContact->name ?? 'Default Contact'}}</label>
                                     </div>
-                                    
-                                    <div class="content ml-3">
-                                        <span class="forum-title text-primary font-weight-bold">{{$myContact->phone_number}}</span>
-                                        <p class="text-muted small mb-0 mt-2">{{$myContact->address}}</p>
-                                    </div>
-
                                 </div>
-                            </td>
-                        </tr>
+                                
+                                <div class="content ml-3">
+                                    <span class="forum-title text-primary font-weight-bold">{{$myContact->phone_number}}</span>
+                                    <p class="text-muted small mb-0 mt-2">{{$myContact->address}}</p>
+                                </div>
+
+                            </div>
+                        </td>
+                    </tr>
                 @endforeach 
             @endif
         </tbody>
