+++ resources/views/client/services/_newAddress.blade.php
@@ -15,14 +15,14 @@
                                     <div class="col-md-4">
                                         <div class="form-group position-relative">
                                             <label>First Name <span class="text-danger">*</span></label>
-                                            <input id="first-name" type="text" class="form-control" placeholder="First Name :" autocomplete="off" />
+                                            <input id="first-name" type="text" class="form-control" placeholder="First Name" autocomplete="off" />
                                         </div>
                                     </div>
                                     <!--end col-->
                                     <div class="col-md-4">
                                         <div class="form-group position-relative">
                                             <label>Last Name <span class="text-danger">*</span></label>
-                                            <input id="last-name" type="text" class="form-control" placeholder="Last Name :" autocomplete="off" />
+                                            <input id="last-name" type="text" class="form-control" placeholder="Last Name" autocomplete="off" />
                                         </div>
                                     </div>
 
@@ -37,8 +37,9 @@
                                     <div class="col-md-4">
                                         <div class="form-group position-relative">
                                             <label>State <span class="text-danger">*</span></label>
-                                            <select class="form-control pl-5 @error('state_id') is-invalid @enderror" id="state_id">
-                                                <option selected value="">Select...</option>
+                                            <select class="form-control @error('state_id') is-invalid @enderror" id="state_id">
+                                                <option selected value="" disabled>Select...</option>
+                                                {{-- <option value="24">Lagos</option> --}}
                                                 @foreach($states as $state)
                                                 <option value="{{ $state->id }}">{{ $state->name }}</option>
                                                 @endforeach
@@ -55,7 +56,7 @@
                                     <div class="col-md-4">
                                         <div class="form-group position-relative">
                                             <label>LGA <span class="text-danger">*</span></label>
-                                            <select class="form-control pl-5 @error('lga_id') is-invalid @enderror" id="lga_id">
+                                            <select class="form-control @error('lga_id') is-invalid @enderror" id="lga_id">
                                                 <option selected value="">Select...</option>
                                             </select>
                                             @error('lga_id')
@@ -71,7 +72,7 @@
                                         <div class="form-group position-relative">
                                             <label>Town/City <span class="text-danger">*</span></label>
 
-                                            <select class="form-control pl-5 @error('town_id') is-invalid @enderror" id="town_id">
+                                            <select class="form-control @error('town_id') is-invalid @enderror" id="town_id">
                                                 <option selected value="">Select...</option>
                                             </select>
                                             @error('town_id')
@@ -79,8 +80,6 @@
                                                 <strong>{{ $message }}</strong>
                                             </span>
                                             @enderror
-
-                                            <!-- <input type="text" name="state" id="state" class="form-control" placeholder="State Name :"> -->
                                         </div>
                                     </div>
 
@@ -91,7 +90,7 @@
                                     <div class="col-12">
                                         <div class="form-group position-relative">
                                             <label>Full Address <span class="text-danger">*</span></label>
-                                            <input type="text" id="street-address" class="form-control user_address" placeholder="Full address of contact :" autocomplete="off" >
+                                            <input type="text" id="street-address" class="form-control user_address" placeholder="Full address of contact" autocomplete="off" >
                                         </div>
                                     </div>
                                     <!--end col-->
